+++ src/tests/EntityGraphQL.Tests/SchemaTests/GraphQLSchemaGenerateTests.cs
@@ -149,14 +149,14 @@ namespace EntityGraphQL.Tests
         [Fact]
         public void TestIgnoreWithSchemaBuilder()
         {
-            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(new SchemaBuilderOptions() { IgnoreTypes = new[] { typeof(Album).FullName }.ToHashSet() });
+            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(new SchemaBuilderOptions() { IgnoreTypes = new[] { typeof(Album) }.ToHashSet() });
             var schema = schemaProvider.ToGraphQLSchemaString();
             Assert.DoesNotContain("album", schema);
         }
         [Fact]
         public void TestIgnoreEnumWithSchemaBuilder()
         {
-            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(new SchemaBuilderOptions() { IgnoreTypes = new[] { typeof(Genre).FullName }.ToHashSet() });
+            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(new SchemaBuilderOptions() { IgnoreTypes = new[] { typeof(Genre) }.ToHashSet() });
             var schema = schemaProvider.ToGraphQLSchemaString();
             Assert.DoesNotContain("genre", schema);
         }